Connection: Improve keepalive handling
This commit reorganizes the handling of different connection loss
situations. In particular calling close() with an error on a WebSocket
connection will now emit a custom 'error' event on the WebSocket object.
Applications are expected to handle this event to detect connections
being closed due to protocol errors or keepalive timeouts.

One side effect of that is that tcp connection loss is detected faster
with WebSockets.